requirements.txt
flask==3.0.3
gunicorn==21.2.0

playwright==1.45.0
aiofiles==24.1.0
tenacity==8.5.0
lxml==5.3.0

google-cloud-storage==2.18.2
google-cloud-logging==3.11.2
